Made error messages more verbose
While I was using your site to learn about Handlebars, I got an error,
which was printed to the screen.  Then I fixed my code, hit the
“Compile” button again, but nothing happened.  It took me a bit to see
that the error message only gets written to when an error is thrown.
There is no feedback confirming the code is fixed and that’s a bit
confusing.  This request has some changes to make this a little more
verbose.  Thanks for considering.